remote/pyproject.toml
[project]
name = "thinky-remote"
version = "0.1.0"
description = "DigitalOcean sandbox VM manager for Thinky Dev Productivity"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.scripts]
thinky-remote = "thinky_remote.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["thinky_remote*"]

[tool.setuptools.package-data]
thinky_remote = ["templates/*.yaml"]
